chore: change impl, remove file from generation -> InputStream
