--- src/main/java/fr/insee/pogues/transforms/visualize/DDIToFO.java
@@ -1,5 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface DDIToFO extends Transformer { }
