--- src/main/java/fr/insee/pogues/transforms/visualize/DDIToFODT.java
@@ -1,5 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface DDIToFODT extends Transformer { }
