--- src/main/java/fr/insee/pogues/transforms/visualize/DDIToLunaticJSON.java
@@ -1,7 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface DDIToLunaticJSON extends Transformer{
-
-}
