+++ src/main/java/fr/insee/pogues/transforms/visualize/FOToPDF.java
@@ -1,7 +1,4 @@
 package fr.insee.pogues.transforms.visualize;
 
-import fr.insee.pogues.transforms.Transformer;
-
-public interface FOToPDF extends Transformer{
-
+public interface FOToPDF extends ModelTransformer {
 }
