--- src/main/java/fr/insee/pogues/transforms/visualize/LunaticJSONToUriStromaeV2.java
@@ -1,7 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface LunaticJSONToUriStromaeV2 extends Transformer {
-
-}
