--- src/main/java/fr/insee/pogues/transforms/visualize/LunaticJSONToUriStromaeV3.java
@@ -1,7 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface LunaticJSONToUriStromaeV3 extends Transformer {
-
-}
