+++ src/main/java/fr/insee/pogues/transforms/visualize/ModelToUriTransformer.java
@@ -1,0 +1,9 @@
+package fr.insee.pogues.transforms.visualize;
+
+import java.io.InputStream;
+import java.net.URI;
+import java.util.Map;
+
+public interface ModelToUriTransformer {
+	URI transform(InputStream inputStream, Map<String, Object> params, String surveyName) throws Exception;
+}
