+++ src/main/java/fr/insee/pogues/transforms/visualize/ModelTransformer.java
@@ -1,0 +1,11 @@
+package fr.insee.pogues.transforms.visualize;
+
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.Map;
+
+public interface ModelTransformer {
+	ByteArrayOutputStream transform(InputStream inputStream, Map<String, Object> params, String surveyName) throws Exception;
+
+}
