+++ src/main/java/fr/insee/pogues/transforms/visualize/PoguesJSONToPoguesJSONDeref.java
@@ -1,6 +1,4 @@
 package fr.insee.pogues.transforms.visualize;
 
-import fr.insee.pogues.transforms.Transformer;
-
-public interface PoguesJSONToPoguesJSONDeref extends Transformer {
+public interface PoguesJSONToPoguesJSONDeref extends ModelTransformer {
 }
