--- src/main/java/fr/insee/pogues/transforms/visualize/PoguesXMLToDDI.java
@@ -1,8 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-/**
- * Created by I6VWID on 13/01/18.
- */
-public interface PoguesXMLToDDI extends Transformer { }
