+++ src/main/java/fr/insee/pogues/transforms/visualize/PoguesXMLToPoguesJSON.java
@@ -1,5 +1,3 @@
 package fr.insee.pogues.transforms.visualize;
 
-import fr.insee.pogues.transforms.Transformer;
-
-public interface PoguesXMLToPoguesJSON extends Transformer { }
+public interface PoguesXMLToPoguesJSON extends ModelTransformer { }
