+++ src/main/java/fr/insee/pogues/transforms/visualize/PoguesXMLToPoguesJSONImpl.java
@@ -1,5 +1,6 @@
 package fr.insee.pogues.transforms.visualize;
 
+import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.Charset;
@@ -11,43 +12,29 @@ import org.springframework.stereotype.Service;
 
 import fr.insee.pogues.conversion.XMLToJSONTranslator;
 
+import static fr.insee.pogues.transforms.visualize.ModelMapper.inputStream2String;
+import static fr.insee.pogues.transforms.visualize.ModelMapper.string2BOAS;
+
 @Service
 public class PoguesXMLToPoguesJSONImpl implements PoguesXMLToPoguesJSON {
 
     private XMLToJSONTranslator translator = new XMLToJSONTranslator(true);
-//
-//    @PostConstruct
-//    public void onInit() {
-//        System.setProperty("javax.xml.bind.JAXBContextFactory", "org.eclipse.persistence.jaxb.JAXBContextFactory");
-//    }
 
-    public void transform(InputStream input, OutputStream output, Map<String, Object> params, String surveyName) throws Exception {
-        if (null == input) {
+    @Override
+    public ByteArrayOutputStream transform(InputStream inputStream, Map<String, Object> params, String surveyName) throws Exception {
+        if (null == inputStream) {
             throw new NullPointerException("Null input");
         }
-        if (null == output) {
-            throw new NullPointerException("Null output");
-        }
-        byte[] out = transform(input, params, surveyName).getBytes(Charset.forName("UTF-8"));
-        output.write(out, 0, out.length);
+        String inputAsString = inputStream2String(inputStream);
+        String outputAsString = transform(inputAsString);
+        return string2BOAS(outputAsString);
     }
 
-    public String transform(InputStream input, Map<String, Object> params, String surveyName) throws Exception {
-        if (null == input) {
-            throw new NullPointerException("Null input");
-        }
-        return transform(IOUtils.toString(input, StandardCharsets.UTF_8.name()), params, surveyName);
-
-    }
-
-    public String transform(String input, Map<String, Object> params, String surveyName) throws Exception {
+    private String transform(String input) throws Exception {
         if (null == input) {
             throw new NullPointerException("Null input");
         }
         try {
-//            JSONParser parser = new JSONParser();
-//            JSONObject questionnaire = (JSONObject) parser.parse(input);
-//            questionnaire = JSONFunctions.renameQuestionnairePlural(questionnaire);
             return translator.translate(input);
         } catch (Exception e) {
         	e.printStackTrace();
@@ -55,5 +42,6 @@ public class PoguesXMLToPoguesJSONImpl implements PoguesXMLToPoguesJSON {
         }
 
     }
+
 
 }
