--- src/main/java/fr/insee/pogues/transforms/visualize/XFormsToURIStromaeV1.java
@@ -1,6 +1,0 @@
-package fr.insee.pogues.transforms.visualize;
-
-import fr.insee.pogues.transforms.Transformer;
-
-public interface XFormsToURIStromaeV1 extends Transformer{
-}
