+++ src/main/java/fr/insee/pogues/transforms/visualize/eno/DDIToFODT.java
@@ -1,0 +1,5 @@
+package fr.insee.pogues.transforms.visualize.eno;
+
+import fr.insee.pogues.transforms.visualize.ModelTransformer;
+
+public interface DDIToFODT extends ModelTransformer { }
