+++ src/main/java/fr/insee/pogues/transforms/visualize/eno/DDIToLunaticJSON.java
@@ -1,0 +1,7 @@
+package fr.insee.pogues.transforms.visualize.eno;
+
+import fr.insee.pogues.transforms.visualize.ModelTransformer;
+
+public interface DDIToLunaticJSON extends ModelTransformer {
+
+}
