+++ src/main/java/fr/insee/pogues/transforms/visualize/eno/DDIToXForms.java
@@ -1,0 +1,5 @@
+package fr.insee.pogues.transforms.visualize.eno;
+
+import fr.insee.pogues.transforms.visualize.ModelTransformer;
+
+public interface DDIToXForms extends ModelTransformer { }
