+++ src/main/java/fr/insee/pogues/transforms/visualize/eno/PoguesXMLToDDI.java
@@ -1,0 +1,8 @@
+package fr.insee.pogues.transforms.visualize.eno;
+
+import fr.insee.pogues.transforms.visualize.ModelTransformer;
+
+/**
+ * Created by I6VWID on 13/01/18.
+ */
+public interface PoguesXMLToDDI extends ModelTransformer { }
