+++ src/main/java/fr/insee/pogues/transforms/visualize/uri/LunaticJSONToUriStromaeV2.java
@@ -1,0 +1,7 @@
+package fr.insee.pogues.transforms.visualize.uri;
+
+import fr.insee.pogues.transforms.visualize.ModelToUriTransformer;
+
+public interface LunaticJSONToUriStromaeV2 extends ModelToUriTransformer {
+
+}
