+++ src/main/java/fr/insee/pogues/transforms/visualize/uri/LunaticJSONToUriStromaeV3Impl.java
@@ -1,4 +1,4 @@
-package fr.insee.pogues.transforms.visualize;
+package fr.insee.pogues.transforms.visualize.uri;
 
 import fr.insee.pogues.persistence.service.QuestionnairesService;
 import fr.insee.pogues.utils.suggester.SuggesterVisuTreatment;
@@ -11,12 +11,14 @@ import org.springframework.stereotype.Service;
 
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.net.URI;
 import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
 
 @Service
-public class LunaticJSONToUriStromaeV2Impl implements LunaticJSONToUriStromaeV2{
+public class LunaticJSONToUriStromaeV3Impl implements LunaticJSONToUriStromaeV3{
 
 	@Autowired
 	private QuestionnairesService questionnaireService;
@@ -30,58 +32,46 @@ public class LunaticJSONToUriStromaeV2Impl implements LunaticJSONToUriStromaeV2{
 	@Value("${application.scheme}")
 	private String apiScheme;
 
-	@Value("${application.stromaev2.vis.host}")
+	@Value("${application.stromaev3.vis.host}")
 	private String orchestratorHost;
 
-	@Value("${application.stromaev2.vis.path}")
+	@Value("${application.stromaev3.vis.path}")
 	private String visualizePath;
 
-	@Value("${application.stromaev2.vis.queryparams.questionnaire}")
+	@Value("${application.stromaev3.vis.queryparams.questionnaire}")
 	private String queryParamsQuestionnaire;
 
-	@Value("${application.stromaev2.vis.queryparams.nomenclatures}")
+	@Value("${application.stromaev3.vis.queryparams.nomenclatures}")
 	private String queryParamsNomenclatures;
 
 	@Value("${application.api.nomenclatures}")
 	private String apiNomenclatures;
 
 	@Override
-	public void transform(InputStream input, OutputStream output, Map<String, Object> params, String surveyName)
-			throws Exception {
-		throw new RuntimeException("Not Implemented");
-
-	}
-
-	@Override
-	public String transform(InputStream input, Map<String, Object> params, String surveyName) throws Exception {
-		throw new RuntimeException("Not Implemented");
-	}
-
-	@Override
-	public String transform(String input, Map<String, Object> params, String surveyName) throws Exception {
+	public URI transform(InputStream input, Map<String, Object> params, String surveyName) throws Exception {
 		JSONParser parser = new JSONParser();
-		JSONObject jsonContent = (JSONObject) parser.parse(input);
-		String id = (String) jsonContent.get("id");
+		JSONObject jsonContent = (JSONObject) parser.parse(new String(input.readAllBytes(), StandardCharsets.UTF_8));
+		String id  = (String) jsonContent.get("id");
 		try {
 			questionnaireService.createJsonLunatic(jsonContent);
 		} catch (PoguesException e) {
 			questionnaireService.updateJsonLunatic(id, jsonContent);
-		} catch (Exception e) {
-			throw new Exception(String.format("%s:%s", getClass().getName(), e.getMessage()));
-		}
+        } catch (Exception e) {
+            throw new Exception(String.format("%s:%s", getClass().getName(), e.getMessage()));
+        }
 
 		List<String> nomenclatureIds = (List<String>) params.get("nomenclatureIds");
 		String jsonStringNomenclaturesForVisu = SuggesterVisuTreatment.createJsonNomenclaturesForVisu(nomenclatureIds, apiNomenclatures).toJSONString();
 		String urlGetJsonLunatic = String.format("%s://%s%s/api/persistence/questionnaire/json-lunatic/%s", apiScheme, apiHost, apiName, id);
 
-		return String.format(
+		return URI.create(String.format(
 				"%s%s?%s=%s&%s=%s",
 				orchestratorHost,
 				visualizePath,
 				queryParamsQuestionnaire,
 				URLEncoder.encode(urlGetJsonLunatic, "UTF-8"),
 				queryParamsNomenclatures,
-				URLEncoder.encode(jsonStringNomenclaturesForVisu, "UTF-8")
+				URLEncoder.encode(jsonStringNomenclaturesForVisu, "UTF-8"))
 		);
 	}
