+++ src/main/java/fr/insee/pogues/transforms/visualize/uri/XFormsToURIStromaeV1.java
@@ -1,0 +1,6 @@
+package fr.insee.pogues.transforms.visualize.uri;
+
+import fr.insee.pogues.transforms.visualize.ModelToUriTransformer;
+
+public interface XFormsToURIStromaeV1 extends ModelToUriTransformer {
+}
